# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Minimal HTTP API that forwards selected Green-API calls. Instance credentials are accepted via `X-Instance-Id` and `X-Api-Token` headers (and may be merged from JSON on POST bodies). Success responses wrap upstream JSON as pretty-printed text in `pretty`.",
        "title": "Green-API HTTP proxy",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Simple browser UI for trying the proxy (not required for API clients).",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "ui"
                ],
                "summary": "Demo UI",
                "responses": {
                    "200": {
                        "description": "HTML page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/get-settings": {
            "get": {
                "description": "Forwards to Green-API using `idInstance` / `apiTokenInstance` from headers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Get instance settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Green-API idInstance (digits)",
                        "name": "X-Instance-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Green-API apiTokenInstance",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok=true, pretty=formatted upstream JSON",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "400": {
                        "description": "validation_error or invalid_json",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "502": {
                        "description": "upstream integration error",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/get-state-instance": {
            "get": {
                "description": "Returns authorization / instance state from Green-API.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Get instance state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Green-API idInstance (digits)",
                        "name": "X-Instance-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Green-API apiTokenInstance",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/send-file-by-url": {
            "post": {
                "description": "JSON body: chatId, fileUrl, fileName; optional caption. Headers override embedded credentials.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Send file by URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Green-API idInstance",
                        "name": "X-Instance-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Green-API apiTokenInstance",
                        "name": "X-Api-Token",
                        "in": "header"
                    },
                    {
                        "description": "Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.SendFileByURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/send-message": {
            "post": {
                "description": "JSON body must include chatId and message. Credentials in headers override body fields when both are present.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Send text message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Green-API idInstance (optional if present in body)",
                        "name": "X-Instance-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Green-API apiTokenInstance (optional if present in body)",
                        "name": "X-Api-Token",
                        "in": "header"
                    },
                    {
                        "description": "Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/green-api-test_internal_domain.APIResponse"
                        }
                    }
                }
            }
        },
        "/livez": {
            "get": {
                "description": "Returns plain text `ok`.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Returns plain text `ok` when startup wiring completed successfully.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "green-api-test_internal_domain.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "green-api-test_internal_domain.APIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/green-api-test_internal_domain.APIError"
                },
                "ok": {
                    "type": "boolean"
                },
                "pretty": {
                    "type": "string"
                }
            }
        },
        "green-api-test_internal_domain.SendFileByURLRequest": {
            "type": "object",
            "properties": {
                "apiTokenInstance": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                },
                "chatId": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                },
                "idInstance": {
                    "type": "string"
                }
            }
        },
        "green-api-test_internal_domain.SendMessageRequest": {
            "type": "object",
            "properties": {
                "apiTokenInstance": {
                    "type": "string"
                },
                "chatId": {
                    "type": "string"
                },
                "idInstance": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}